+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### :bug: Fixed
 
 - False positive `negative_data_rejection` for boolean query/path parameters in the coverage phase. [#4254](https://github.com/schemathesis/schemathesis/issues/4254)
+- False positive `negative_data_rejection` when an API key is sent as a query parameter. [#4254](https://github.com/schemathesis/schemathesis/issues/4254)
 
 ## [4.21.9](https://github.com/schemathesis/schemathesis/compare/v4.21.8...v4.21.9) - 2026-06-19